refactor(orders): wave-3 slice 11 work-completion scenario tables (#3949)
Migrate four imperative work-completion/dispatch suites into one thin
family runner with support scenario/expectation tables.
